UPDATE: simplify app structure
