+++ ai/gamesimulation.py
@@ -1,6 +1,144 @@
-import random, math
-from .game import AI_ball, HEIGHT, WIDTH, BALL_SPEED, PADDLE_WIDTH, PADDLE_HEIGHT, BALL_SIZE, PADDLE_SPEED, GRID
+import random, math, pygame
 from .gameconfig import get_game_config
+
+# Set up the game window
+WIDTH = 80 * 6
+HEIGHT = 24 * 10
+GRID = 5
+
+# Colors
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
+RED = (255, 0, 0)
+
+# Paddle settings
+PADDLE_HEIGHT = 5 * 6
+PADDLE_WIDTH = 5
+PADDLE_SPEED = 2
+
+# Ball settings
+BALL_SPEED = 2
+BALL_SIZE = 5
+BALL_MIN_DY = 0.5
+ball : pygame.rect
+ball_dx : float
+ball_dy : float
+
+class AI_ball:
+    x: float
+    y: float
+    dx: float
+    dy: float
+
+    def __init__(self, x, y=None, dx=None, dy=None):
+        if y == None:
+            ball = x
+            self.update(ball, ball_dx, ball_dy)
+        else:
+            self.x = x
+            self.y = y
+            self.dx = dx
+            self.dy = dy
+
+    def __repr__(self):
+        return f"x = {self.x}\t\t\ty = {self.y} \ndx = {self.dx}\t\t\tdy = {self.dy}\n"
+    
+    def update(self, ball, ball_dx, ball_dy):
+        self.x = ball.x
+        self.y = ball.y
+        self.dx = ball_dx
+        self.dy = ball_dy
+
+# Create paddles and ball
+class Paddle:
+    def __init__(self, x, y, width, height):
+        self.x = x
+        self.y = y
+        self.width = width
+        self.height = height
+        # Add center property to match pygame.Rect behavior
+        self._center = (x + width/2, y + height/2)
+    
+    @property
+    def center(self):
+        return self._center
+    
+    @center.setter 
+    def center(self, value):
+        self._center = value
+        self.x = value[0] - self.width/2
+        self.y = value[1] - self.height/2
+    
+    @property
+    def top(self):
+        return self.y
+    
+    @property
+    def bottom(self):
+        return self.y + self.height
+    
+    @property
+    def left(self):
+        return self.x
+    
+    @property
+    def right(self):
+        return self.x + self.width
+
+class Ball:
+    def __init__(self, x, y, size):
+        self.x = x
+        self.y = y
+        self.size = size
+        self.width = size
+        self.height = size
+        self._center = (x, y)
+    
+    @property
+    def center(self):
+        return self._center
+    
+    @center.setter
+    def center(self, value):
+        self._center = value
+        self.x = value[0]
+        self.y = value[1]
+    
+    @property 
+    def top(self):
+        return self.y - self.size/2
+    
+    @top.setter
+    def top(self, value):
+        self.y = value + self.size/2
+        self._center = (self.x, self.y)
+    
+    @property
+    def bottom(self):
+        return self.y + self.size/2
+    
+    @bottom.setter
+    def bottom(self, value):
+        self.y = value - self.size/2
+        self._center = (self.x, self.y)
+    
+    @property
+    def left(self):
+        return self.x - self.size/2
+    
+    @left.setter
+    def left(self, value):
+        self.x = value + self.size/2
+        self._center = (self.x, self.y)
+    
+    @property
+    def right(self):
+        return self.x + self.size/2
+    
+    @right.setter
+    def right(self, value):
+        self.x = value - self.size/2
+        self._center = (self.x, self.y)
 
 def ai_bonus_score(ball_y, rightPaddle, Ai_selected):
     dist = abs(ball_y - rightPaddle.top) / HEIGHT
@@ -33,98 +171,11 @@ def updateBallAngle(ball, ball_dy, paddle):
 def generate_random_number(low, high):
     return random.randint(low, high)
 
-def train_basic_no_display(Ai_selected):
-    # Create paddles and ball
-    class Paddle:
-        def __init__(self, x, y, width, height):
-            self.x = x
-            self.y = y
-            self.width = width
-            self.height = height
-            # Add center property to match pygame.Rect behavior
-            self._center = (x + width/2, y + height/2)
-        
-        @property
-        def center(self):
-            return self._center
-        
-        @center.setter 
-        def center(self, value):
-            self._center = value
-            self.x = value[0] - self.width/2
-            self.y = value[1] - self.height/2
-        
-        @property
-        def top(self):
-            return self.y
-        
-        @property
-        def bottom(self):
-            return self.y + self.height
-        
-        @property
-        def left(self):
-            return self.x
-        
-        @property
-        def right(self):
-            return self.x + self.width
+def train_basic(Ai_selected):
+    # train_predefined(Ai_selected)
+    train_normal(Ai_selected)
 
-    class Ball:
-        def __init__(self, x, y, size):
-            self.x = x
-            self.y = y
-            self.size = size
-            self.width = size
-            self.height = size
-            self._center = (x, y)
-        
-        @property
-        def center(self):
-            return self._center
-        
-        @center.setter
-        def center(self, value):
-            self._center = value
-            self.x = value[0]
-            self.y = value[1]
-        
-        @property 
-        def top(self):
-            return self.y - self.size/2
-        
-        @top.setter
-        def top(self, value):
-            self.y = value + self.size/2
-            self._center = (self.x, self.y)
-        
-        @property
-        def bottom(self):
-            return self.y + self.size/2
-        
-        @bottom.setter
-        def bottom(self, value):
-            self.y = value - self.size/2
-            self._center = (self.x, self.y)
-        
-        @property
-        def left(self):
-            return self.x - self.size/2
-        
-        @left.setter
-        def left(self, value):
-            self.x = value + self.size/2
-            self._center = (self.x, self.y)
-        
-        @property
-        def right(self):
-            return self.x + self.size/2
-        
-        @right.setter
-        def right(self, value):
-            self.x = value - self.size/2
-            self._center = (self.x, self.y)
-
+def train_predefined(Ai_selected):
     # Initialize game objects
     rightPaddle = Paddle(
         x = WIDTH - 50 - PADDLE_WIDTH,
@@ -133,72 +184,81 @@ def train_basic_no_display(Ai_selected):
         height = PADDLE_HEIGHT
     )
 
-    # # Predifine game loop
-    # total_predefined = 0
-    # for predefined_y in range(round(HEIGHT / 10), HEIGHT, round(HEIGHT / 10)):
-    #     ball = Ball(
-    #         x = 50,
-    #         y = predefined_y,
-    #         size = BALL_SIZE
-    #     )
+    # Predifine game loop
+    total_predefined = 0
+    for predefined_y in range(round(HEIGHT / 10), HEIGHT, round(HEIGHT / 10)):
+        ball = Ball(
+            x = 50,
+            y = predefined_y,
+            size = BALL_SIZE
+        )
 
-    #     for predefined_angle in range(-75, 76):
-    #         total_predefined += 1
-    #         # Ball movement
-    #         ball_dx = BALL_SPEED
-    #         ball_dy = BALL_SPEED * -math.sin(predefined_angle)
+        for predefined_angle in range(-75, 76):
+            total_predefined += 1
+            # Ball movement
+            ball_dx = BALL_SPEED
+            ball_dy = BALL_SPEED * -math.sin(predefined_angle)
 
-    #         # Update AI's target position
-    #         ai_ball = AI_ball(ball, 0, 0, 0)
-    #         ai_ball.update(ball, ball_dx, ball_dy)
+            # Update AI's target position
+            ai_ball = AI_ball(ball, 0, 0, 0)
+            ai_ball.update(ball, ball_dx, ball_dy)
 
-    #         i = 0
-    #         while (True):
-    #             # Move the ball
-    #             ball.x += ball_dx
-    #             ball.y += ball_dy
+            i = 0
+            while (True):
+                # Move the ball
+                ball.x += ball_dx
+                ball.y += ball_dy
 
-    #             # Update the ai view
-    #             if i % 60 == 0:
-    #                 ai_ball.update(ball, ball_dx, ball_dy)
+                # Update the ai view
+                if i % 60 == 0:
+                    ai_ball.update(ball, ball_dx, ball_dy)
 
-    #             # Move the right paddle
-    #             match (Ai_selected.decision(rightPaddle.y, ai_ball, HEIGHT)):
-    #                 case 0:
-    #                     if rightPaddle.top > 0:
-    #                         rightPaddle.y -= PADDLE_SPEED
-    #                 case 1:
-    #                     pass
-    #                 case 2:
-    #                     if rightPaddle.bottom < HEIGHT:
-    #                         rightPaddle.y += PADDLE_SPEED
+                # Move the right paddle
+                match (Ai_selected.decision(rightPaddle.y, ai_ball, HEIGHT)):
+                    case 0:
+                        if rightPaddle.top > 0:
+                            rightPaddle.y -= PADDLE_SPEED
+                    case 1:
+                        pass
+                    case 2:
+                        if rightPaddle.bottom < HEIGHT:
+                            rightPaddle.y += PADDLE_SPEED
 
-    #             # Ball collision with top and bottom
-    #             if ball.top <= 0:
-    #                 ball.top = GRID
-    #                 ball_dy *= -1
-    #                 if abs(ball_dy) < 1:
-    #                     ball_dy = 1 if ball_dy > 0 else -1
-    #             elif ball.bottom >= HEIGHT:
-    #                 ball.bottom = HEIGHT - GRID
-    #                 ball_dy *= -1
-    #                 if abs(ball_dy) < 1:
-    #                     ball_dy = 1 if ball_dy > 0 else -1
+                # Ball collision with top and bottom
+                if ball.top <= 0:
+                    ball.top = GRID
+                    ball_dy *= -1
+                    if abs(ball_dy) < 1:
+                        ball_dy = 1 if ball_dy > 0 else -1
+                elif ball.bottom >= HEIGHT:
+                    ball.bottom = HEIGHT - GRID
+                    ball_dy *= -1
+                    if abs(ball_dy) < 1:
+                        ball_dy = 1 if ball_dy > 0 else -1
 
-    #             # Ball collision with AI's paddles
-    #             if collides(ball, rightPaddle):
-    #                 if ball.right < rightPaddle.right:
-    #                     Ai_selected.ai_score += 1
-    #                     break
+                # Ball collision with AI's paddles
+                if collides(ball, rightPaddle):
+                    if ball.right < rightPaddle.right:
+                        Ai_selected.ai_score += 1
+                        break
 
-    #             # Ball out of bounds
-    #             if ball.right >= WIDTH:
-    #                 ai_bonus_score(ball.y, rightPaddle, Ai_selected)
-    #                 break
+                # Ball out of bounds
+                if ball.right >= WIDTH:
+                    ai_bonus_score(ball.y, rightPaddle, Ai_selected)
+                    break
                 
-    #             i += 1
+                i += 1
 
-    # print(f"\nScore before normal game: {Ai_selected.ai_score:.1f} / {total_predefined}")
+    print(f"\nScore before normal game: {Ai_selected.ai_score:.1f} / {total_predefined}")
+
+def train_normal(Ai_selected):
+    # Initialize game objects
+    rightPaddle = Paddle(
+        x = WIDTH - 50 - PADDLE_WIDTH,
+        y = HEIGHT//2 - PADDLE_HEIGHT//2,
+        width = PADDLE_WIDTH,
+        height = PADDLE_HEIGHT
+    )
 
     # Normal game loop
     ball = Ball(
@@ -285,4 +345,3 @@ def train_basic_no_display(Ai_selected):
         # End the game
         if left_score >= get_game_config('max_score')[0]:
             running = False
-
